% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSugar.R
\name{find_batch_id}
\alias{find_batch_id}
\title{Find the id of a batch given the name}
\usage{
find_batch_id(project.name, batch.name)
}
\arguments{
\item{project.name}{The project name}

\item{batch.name}{The batch name}
}
\description{
Find the id of a batch given the name
}
