% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSugar.R
\name{find_project_id}
\alias{find_project_id}
\title{Find the id of a project given the name}
\usage{
find_project_id(project.name)
}
\arguments{
\item{project.name}{The project name}
}
\description{
Find the id of a project given the name
}
