% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSugar.R
\name{find_sheet_id}
\alias{find_sheet_id}
\title{Find the id of a sheet given the name}
\usage{
find_sheet_id(project.name, batch.name, sheet.name)
}
\arguments{
\item{project.name}{The project name}

\item{batch.name}{The batch name}

\item{sheet.name}{The sheet name}
}
\description{
Find the id of a sheet given the name
}
