% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSugar.R
\name{get_batch_files}
\alias{get_batch_files}
\title{Return a list of files in a batch as a dataframe}
\usage{
get_batch_files(batch.id)
}
\arguments{
\item{batch.id}{The id of the batch}
}
\description{
Return a list of files in a batch as a dataframe
}
