% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSugar.R
\name{get_batch_files_by_name}
\alias{get_batch_files_by_name}
\title{Return a list of files in a batch as a dataframe}
\usage{
get_batch_files_by_name(project.name, batch.name)
}
\arguments{
\item{project.name}{The project name}

\item{batch.name}{The batch name}
}
\description{
Return a list of files in a batch as a dataframe
}
