% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSugar.R
\name{get_file_contents}
\alias{get_file_contents}
\title{Return the contents of a file given the id}
\usage{
get_file_contents(file.id)
}
\arguments{
\item{file.id}{The id of the sheet}
}
\description{
Return the contents of a file given the id
}
