% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSugar.R
\name{get_sheet_contents}
\alias{get_sheet_contents}
\title{Return the contents of a sheet as a dataframe}
\usage{
get_sheet_contents(sheet.id)
}
\arguments{
\item{sheet.id}{The id of the sheet}
}
\description{
Return the contents of a sheet as a dataframe
}
