% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSugar.R
\name{get_sheet_contents_by_name}
\alias{get_sheet_contents_by_name}
\title{Return the contents of a sheet ass a dataframe, given the project/batch/sheet names.}
\usage{
get_sheet_contents_by_name(project.name, batch.name, sheet.name)
}
\arguments{
\item{project.name}{The project name}

\item{batch.name}{The batch name}

\item{sheet.name}{The sheet name}
}
\description{
Return the contents of a sheet ass a dataframe, given the project/batch/sheet names.
}
